feat: 项目评估

+++ backend/src/services/color_card_scan_service.rs
@@ -65,12 +65,13 @@ impl ColorCardScanService {
 
         // 3. 加载配方（如有）
         let recipe_summary = if let Some(recipe_id) = item.dye_recipe_id {
-            let recipe = crate::models::dye_recipe::Entity::find_by_id(recipe_id)
+            // dye_recipe.id 是 i32，color_card_item.dye_recipe_id 是 Option<i64>，需转换
+            let recipe = crate::models::dye_recipe::Entity::find_by_id(recipe_id as i32)
                 .one(&*self.db)
                 .await
                 .map_err(|e| AppError::database(e.to_string()))?;
             recipe.map(|r| RecipeSummary {
-                id: r.id,
+                id: r.id as i64,
                 recipe_name: r.recipe_name.unwrap_or_default(),
                 fabric_type: r.fabric_type,
                 color_no: r.color_no,
